Move logic to controller
